+++ src/main/java/com/balugaq/constructionwand/core/listeners/PrepareBuildingListener.java
@@ -73,7 +73,7 @@ public class PrepareBuildingListener implements Listener {
             }
         }
 
-        Set<Location> showingBlocks = WandUtil.getBuildingLocations(player, Math.min(limitBlocks, playerHas), buildingWand.getAxis(player.getInventory().getItemInMainHand()), buildingWand.isBlockStrict());
+        Set<Location> showingBlocks = WandUtil.getBuildingLocations(player, Math.min(limitBlocks, playerHas), WandUtil.getAxis(player.getInventory().getItemInMainHand()), buildingWand.isBlockStrict());
         DisplayGroup displayGroup = new DisplayGroup(player.getLocation(), 0.0F, 0.0F);
         for (Location location : showingBlocks) {
             String ls = location.getBlockX() + "_" + location.getBlockY() + "_" + location.getBlockZ();
