--- src/main/java/com/balugaq/constructionwand/utils/Keys.java
@@ -1,11 +1,0 @@
-package com.balugaq.constructionwand.utils;
-
-import lombok.experimental.UtilityClass;
-import org.bukkit.NamespacedKey;
-import org.jetbrains.annotations.NotNull;
-
-@UtilityClass
-public class Keys {
-    public static final NamespacedKey MAIN_GROUP = KeyUtil.newKey("main");
-    public static final NamespacedKey AXIS = KeyUtil.newKey("axis");
-}
